- made a key improvement to the decision process for "nested" cursor: we now check for either the "foundationdb_nested" execution option, or we look in the compiled object to see if nested objects have been encountered.  This has lots of advantages:
1. the behavior of SQL is now more recognizable, as far as when a traditional
column-level subquery is used; we get back a scalar result when nested loading
isn't used

2. the restriction that we can't get back duplicate column names is now lifted
for all SQL except for nested queries.

3. the performance overhead of nested queries, including the nested cursor,
JSON parsing, etc. is now gone for the vast majority of queries.

4. normal usage with Core or ORM constructs selects the "nested" result option
automatically, no extra commands needed, and also virtually no extra
overhead to detect this.

Caveats:

1. a straight textual SQL that wants nested results now needs to flag
the foundationdb_nested option.

2. we might need to add some functionality in the weird case that someone
combines the use of a traditional scalar subquery with a nested() construct
at the same time; the DBAPI will return both of these as "nested" and we may
want to convert one to a scalar (or maybe not).   this is a super edge case.

- updated the README to refer to the new option

- fixed the name of the dialect entrypoint in README, it's "foundationdb+psycopg2",
the "sqlalchemy_" prefix was yet another remnant of the search and replace...